requirements.txt
google-cloud-pubsub
google-auth
configparser
apache-beam[gcp]
google-cloud-bigquery
